Remove non-JSON message Hue broker message
In Hue discovery, if no devices are found via uPnP, the Hue broker service is used. Under normal circumstances, this will return a JSON array containing bridge data. Sometimes, however, non-JSON data is returned. Originally, this was output to the console to find out what it was as this behavior is undocumented. Now that it is know that the none JSON data is HTML saying that the broker service is temporarily unavailable, the message is no longer necessary.