Fix LIFX may crash one message spike
In order to map request packets to response packets, LIFX has a built-in sequence id that goes into the packet. This id is one byte which gives us 255 ids before we must loop around and reuse old ids (id 0 is for non-sequenced packets). Normally, this is fine as the mechanism storing the callback for the request cleans itself up once the response happens. If, however, the ids loop around between the request and response, there may be a new type of callback queued for that sequence id. As such, this may occur if we have over 255 messages within the couple of milliseconds it typically takes between request and response and the newly queued callback is of a different type than the original queued item. To fix, changed the sequencer to be able to queue multiple callbacks for different message types on the same sequence id.